MTD: bcm63xxpart: Recognize alternate tags
Many bcm63xx systems implement a "ping-pong" scheme that allows for
failsafe firmware upgrades.  If both the new image and the old image are
smaller than FLASH_SIZE / 2, CFE will write the new image in the opposite
half of the flash device.  A sequence number is added to the bcm_tag in
order to track which image is most recent.

This patch causes bcm63xxpart to check the sequence numbers when
appropriate, and point the kernel/rootfs partitions to the "alternate"
set when they are active.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>