MIPS: BCM63XX: Clarify HSSPI constants
The TX/RX buffers are asymmetric on the HSSPI controller.  The maximum
length of a TX transaction is 510 bytes, because every transaction
requires a 2-byte opcode.

For simplicity we will advertise our maximum transfer size as 510 bytes,
as there is currently no way to tell m25p80 about the nuances of the
controller (asymmetry, shared RAM vs dedicated TX/RX RAMs, etc.).

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>