MIPS: BCM63XX: nvram: New API to return the boot partition setting
If two images are present in the flash, CFE can either boot the old image
(lower sequence number) or the new image (higher sequence number):

On CFE 4.06L and lower, this was determined by the "p=0" or "p=1" item in
the nvram boot_info string in flash.

On CFE 4.08L and higher, this is passed to Linux through the new "blparms"
struct in RAM (falling back to nvram only if necessary).

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>