bcm63xx_enet: move core reset from clock to driver
Instead of hiding any reset requirements in the clock driver, make it
explicit in the actual ethernet driver.

So far we only need to know if it is BCM6368 or not, so no need to add
entries for each supported SoC. If we need further destinction, we can
add more entries.

Signed-off-by: Jonas Gorski <jonas.gorski@gmail.com>