b53: fix flushing arl entries on BCM5389
BCM5389 uses a slightly different format for configuring filters for the
arl flush, its PORT and VID registers have a special "ALL" value instead
of bits in the FAST_AGE_CTRL register.

Besides that, it also seems to always flush dynamic entries, and the
only defined bit other than the START/DONE one is the one for static
entries.

So make sure we setup the registers correctly for BCM5389 and do not
modify reserved bits.

Fixes: a95691bc54af ("net: dsa: b53: Add BCM5389 support")
Signed-off-by: Jonas Gorski <jonas.gorski@gmail.com>