net: dsa: b53: invert bcm531x5 rgmii delay heuristic
When configuring the RGMII port for BCM531x5, we are only doing so in
MLO_AN_FIXED, i.e. fixed link mode. This implies this port is used as
cpu port and thus connected to an ethernet mac, so the switch takes the
role of the PHY.

Therefore we need to enable delays when the interface mode is rgmii-id,
and not rgmii. Enabling delays for rgmii is wrong in any case.

Also update the comment with the incomplete sentence to match what we
are now doing, and add a comment about rgmii-id.

Luckily there are at least no in-tree users of rgmii or rgmii-id mode
that would be affected, as the only user with a defined link mode uses
rgmii-txid, which we already handle correctly.

Fixes: 967dd82ffc52 ("net: dsa: b53: Add support for Broadcom RoboSwitch")
Reported-by: Álvaro Fernández Rojas <noltari@gmail.com>
Signed-off-by: Jonas Gorski <jonas.gorski@gmail.com>