net: dsa: b53: implement "slow" fast ageing
Implement "slow" fast ageing by iterating over the fdb and removing all
matching entries, and use it for BCM5325, which does not implement fast
ageing.

To also allow removing entries on any port, allow passing -1 as port to
b53_fdb_dump() as "any", and only check the port before calling the cb()
if the port is not negative.

It does not matter that the cb() does not tell us the port of the found
entry, as for (unicast) entry invalidation the port does not matter.

Signed-off-by: Jonas Gorski <jonas.gorski@gmail.com>